Hotfix recent events not showing up.
For chapters which have more events than the default page size of the
Meetup Events API (somewhere in the 600s range), only old events show
up.  This is due to the chronological default ordering of the results.
By reversing the order we still lose some results in the overall counts
but at least get the recency right.